backend/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
click==8.1.7
dnspython==2.6.1
fastapi==0.112.2
h11==0.14.0
httptools==0.6.1
idna==3.8
motor==3.5.1
passlib==1.7.4
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.9.0
pymongo==4.8.0
python-dotenv==1.0.1
PyYAML==6.0.2
sniffio==1.3.1
starlette==0.38.2
typing_extensions==4.12.2
uvicorn==0.30.6
uvloop==0.20.0
watchfiles==0.23.0
websockets==13.0
python-multipart~=0.0.9
